requirements.txt
asgiref==3.3.4
attrs==20.3.0
bitmex==0.2.2
bravado==11.0.3
bravado-core==5.17.0
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-extensions==3.1.3
django-model-utils==4.1.1
django-shell-plus==1.1.7
djangorestframework==3.12.4
idna==2.10
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
monotonic==1.6
msgpack==1.0.2
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3987==1.3.8
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
typing-extensions==3.7.4.3
urllib3==1.26.4
webcolors==1.11.1
